requirements.txt
Flask==2.0.1
flask-cors==3.0.10
mysql-connector-python==8.0.26
Werkzeug==2.0.1
Pillow==9.5.0
fpdf==1.7.2
opencv-python==4.8.1.78